server/requirements.txt
fastapi==0.88.0
pydantic==1.10.2
pymongo==4.3.3
python-decouple==3.6
bcrypt==4.0.1
